Refactor, move, and split fixtures into separate directory.
