+++ test/browser/conftest.py
@@ -1,7 +1,7 @@
-from _helper.fixture import browser_default  # noqa
-from _helper.fixture import browser_default_disable_images  # noqa
-from _helper.fixture import browser_default_headless  # noqa
-from _helper.fixture import browser_default_headless_images_disabled  # noqa
-from _helper.fixture import browser_default_headless_scope_function  # noqa
-from _helper.fixture import browser_headless_screenshot  # noqa
-from _helper.fixture import window_handle_controller  # noqa
+from _fixture.browser import browser_default  # noqa
+from _fixture.browser import browser_default_disable_images  # noqa
+from _fixture.browser import browser_default_headless  # noqa
+from _fixture.browser import browser_default_headless_images_disabled  # noqa
+from _fixture.browser import browser_default_headless_scope_function  # noqa
+from _fixture.browser import browser_headless_screenshot  # noqa
+from _fixture.window_handle import window_handle_controller  # noqa
